refactor: separate notification functions
